// client/src/app/shared/models/products.ts
export interface Product {
  id: number
  name: string
  description: string
  price: number
  quantity_In_Stock: number
  origin: string
  type: string
  flavor: string
  caffeineContent: string
  imageUrl: string
  manufacturingDate: string
  expirationDate: string
  steepingTime?: string
  waterTemperature?: string
  color?: string
  caffeineLevel?: string
  healthBenefits: string
  roastLevel?: string
  grindType?: string
  brewingMethod?: string
  productType: string
  productBrand: string
}